+++ App.js
@@ -8,43 +8,39 @@ import {
 import { createStackNavigator } from 'react-navigation-stack'
 
 // AuthStack
-import AuthLoadingScreen from './src/components/auth/AuthLoadingScreen'
 import SignUpScreen from './src/components/auth/SignUpScreen'
 import SignInScreen from './src/components/auth/SignInScreen'
 import ForgotPasswordScreen from './src/components/auth/ForgotPasswordScreen'
 import SettingsScreen from './src/components/auth/SettingsScreen'
 
 // DashboardStack
-import HomeScreen from './src/components/user/HomeScreen'
-import ProfileScreen from './src/components/user/ProfileScreen'
+import UserDashboardScreen from './src/components/user/UserDashboardScreen'
 
-// Debug tool
-import DebugScreen from './src/components/main/DebugScreen'
+// Main stack
+import HomeScreen from './src/components/main/HomeScreen'
 
 // Amplify imports and config
 import Amplify from '@aws-amplify/core'
 import config from './aws-exports'
 Amplify.configure(config)
 
-// Dashboard stack
-const DashboardStackNavigator = createStackNavigator({
-  Home: HomeScreen,
-  Profile: ProfileScreen,
-}, { headerMode: 'none' })
-
 // Auth stack
 const AuthStackNavigator = createStackNavigator({
-  Debug: DebugScreen,
-  SignUp: SignUpScreen,
   SignIn: SignInScreen,
+  SignUp: SignUpScreen,
   ForgotPassword: ForgotPasswordScreen,
-  Settings: SettingsScreen,
+}, { headerMode: 'none' })
+
+// Main stack
+const MainStackNavigator = createStackNavigator({
+  Home: HomeScreen,
+  UserDashboard: UserDashboardScreen, // DashboardStack
+  Settings: SettingsScreen, 
+  Auth: AuthStackNavigator, // AuthStackNavigator
 }, { headerMode: 'none' })
 
 const AppSwitchNavigator = createSwitchNavigator({
-  AuthLoading: AuthLoadingScreen,
-  Auth: AuthStackNavigator, // the AuthStack
-  Dashboard: DashboardStackNavigator // DashboardStack
+  Main: MainStackNavigator, // the MainStack
 })
 
 const App = createAppContainer(AppSwitchNavigator)
